some refactoring

+++ BI/GoogleProvider.cs
@@ -1,5 +1,5 @@
 using Newtonsoft.Json;
-using System.Net;
+using System.Net.Http.Headers;
 using System.Text;
 
 namespace MyDrive;
@@ -8,9 +8,11 @@ public class GoogleProvider
 {
     private List<(string id, GoogleAccess access)> _cache = new();
     private readonly IConfiguration _config;
-    public GoogleProvider(IConfiguration config)
+    private readonly HttpClient _httpClient;
+    public GoogleProvider(IConfiguration config, HttpClient httpClient)
     {
         _config = config;
+        _httpClient = httpClient;
     }
 
     private string GoogleClientId { get => _config.GetGoogleClientId(); }
@@ -30,25 +32,25 @@ public class GoogleProvider
         body.Append("redirect_uri=" + redirect + "&");
         body.Append("grant_type=authorization_code");
         var res = await Post("https://www.googleapis.com/oauth2/v4/token", body.ToString());
-        var bearer = JsonConvert.DeserializeObject<JsonBearerToken>(res.body);
-        var storage = GetConfigStorage(id);
-        await storage.Delete();
+        var bearer = JsonConvert.DeserializeObject<JsonBearerToken>(res);
+        if (bearer == null)
+        {
+            throw new Exception("couldn't get bearer");
+        }
         var config = new JsonGoogleConfig()
         {
             RefreshToken = bearer.refresh_token
         };
-        using (var mem = new MemoryStream())
-        using (var tempWriter = new StreamWriter(mem))
-        {
-            await tempWriter.WriteAsync(JsonConvert.SerializeObject(config));
-            await tempWriter.FlushAsync();
-            using (var uploader = await storage.CreateUploadStream(mem.Length))
-            using (var writer = new StreamWriter(uploader))
-            {
-                await writer.WriteAsync(JsonConvert.SerializeObject(config));
-                await writer.FlushAsync();
-            }
-        }
+        var storage = GetConfigStorage(id);
+        await storage.Delete();
+        using var mem = new MemoryStream();
+        using var tempWriter = new StreamWriter(mem);
+        await tempWriter.WriteAsync(JsonConvert.SerializeObject(config));
+        await tempWriter.FlushAsync();
+        using var uploader = await storage.CreateUploadStream(mem.Length);
+        using var writer = new StreamWriter(uploader);
+        await writer.WriteAsync(JsonConvert.SerializeObject(config));
+        await writer.FlushAsync();
     }
 
     public async Task<GoogleAccess?> GetAccess(string id)
@@ -64,8 +66,12 @@ public class GoogleProvider
             return null;
         }
         var refresh = JsonConvert.DeserializeObject<JsonGoogleConfig>(refreshToken)?.RefreshToken;
-        Func<Task<string?>> func = new AccessTokenProvider(refresh, FetchAccessTokenFromGoogle).GetAccessToken;
-        var access = new GoogleAccess(func);
+        if (refresh == null)
+        {
+            return null;
+        }
+        Func<Task<string?>> getAccessToken = new AccessTokenProvider(refresh, FetchAccessTokenFromGoogle).GetAccessToken;
+        var access = new GoogleAccess(getAccessToken);
         if (await access.HasAccess())
         {
             _cache.Add((id, access));
@@ -92,11 +98,11 @@ public class GoogleProvider
         sb.Append("response_type=code&");
         sb.Append("login_hint=" + email + "&");
         sb.Append("prompt=select_account&");
-        sb.Append("client_id=" + GoogleClientId + "&");
+        sb.Append("client_id=" + GoogleClientId + "&"); // still needs to add redirect_url, but the clients adds it per its host
         return sb.ToString();
     }
 
-    private async Task<JsonAccessToken> FetchAccessTokenFromGoogle(string refreshToken)
+    private async Task<JsonAccessToken?> FetchAccessTokenFromGoogle(string refreshToken)
     {
         Console.WriteLine("FetchAccessTokenFromGoogle");
         if (string.IsNullOrEmpty(refreshToken))
@@ -111,7 +117,7 @@ public class GoogleProvider
             body.Append("refresh_token=" + refreshToken + "&");
             body.Append("grant_type=refresh_token");
             var res = await Post("https://www.googleapis.com/oauth2/v4/token", body.ToString());
-            var access = JsonConvert.DeserializeObject<JsonAccessToken>(res.body);
+            var access = JsonConvert.DeserializeObject<JsonAccessToken>(res);
             return access;
         }
         catch
@@ -120,29 +126,14 @@ public class GoogleProvider
         }
     }
 
-    private async Task<(WebResponse response, string body)> Post(string uri, string body)
+    private async Task<string> Post(string uri, string body)
     {
-        HttpWebRequest request = (HttpWebRequest)WebRequest.Create(uri);
-        request.AutomaticDecompression = DecompressionMethods.GZip | DecompressionMethods.Deflate;
-        request.Method = "POST";
-        var data = Encoding.ASCII.GetBytes(body);
-        request.ContentType = "application/x-www-form-urlencoded";
-        request.ContentLength = data.Length;
-        using (var stream = await request.GetRequestStreamAsync())
-        {
-            stream.Write(data, 0, data.Length);
-        }
-        try
-        {
-            var response = (HttpWebResponse)await request.GetResponseAsync();
-            string responseString = await new StreamReader(response.GetResponseStream()).ReadToEndAsync();
-            return (response, responseString);
-        }
-        catch (Exception ex)
-        {
-            Console.WriteLine(ex.Message);
-            throw;
-        }
+        using var request = new HttpRequestMessage(HttpMethod.Post, uri);
+        request.Content = new ByteArrayContent(Encoding.ASCII.GetBytes(body));
+        request.Content.Headers.ContentType = new MediaTypeHeaderValue("application/x-www-form-urlencoded");
+        using HttpResponseMessage response = await _httpClient.SendAsync(request).ConfigureAwait(false);
+        var ans = await response.Content.ReadAsStringAsync().ConfigureAwait(false);
+        return ans;
     }
 
     private class JsonAccessToken
@@ -170,12 +161,12 @@ public class GoogleProvider
     private class AccessTokenProvider
     {
         private string _refreshToken;
-        private readonly Func<string, Task<JsonAccessToken>> _fetchAccessToken;
+        private readonly Func<string, Task<JsonAccessToken?>> _fetchAccessToken;
 
         private string? AccessToken { get; set; } = null;
         private DateTime AccessTokenExpiration { get; set; } = DateTime.MinValue;
 
-        public AccessTokenProvider(string refreshToken, Func<string, Task<JsonAccessToken>> fetchAccessToken)
+        public AccessTokenProvider(string refreshToken, Func<string, Task<JsonAccessToken?>> fetchAccessToken)
         {
             this._refreshToken = refreshToken;
             this._fetchAccessToken = fetchAccessToken;
@@ -191,11 +182,13 @@ public class GoogleProvider
             if (_refreshToken != null)
             {
                 var access = await _fetchAccessToken(_refreshToken);
-                this.AccessToken = access.access_token;
-                this.AccessTokenExpiration = DateTime.UtcNow + TimeSpan.FromSeconds(access.expires_in - 10);
-                return AccessToken;
+                if (access != null)
+                {
+                    this.AccessToken = access.access_token;
+                    this.AccessTokenExpiration = DateTime.UtcNow + TimeSpan.FromSeconds(access.expires_in - 10);
+                    return AccessToken;
+                }
             }
-
             return null;
         }
     }
